backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic[email]==1.10.9
python-multipart==0.0.20
sqlalchemy
psycopg[binary]
gunicorn
python-dotenv
asyncpg